move DateValue to util package
